final edit 3

+++ Core/Inc/max9812.h
@@ -3,5 +3,5 @@
 double VoltsToDecibels();
 uint32_t read_analog();
 
-extern int signalMax;
-extern int signalMin;
+extern uint16_t signalMax;
+extern uint16_t signalMin;
